implemented scoring in AddEnergyScreen
