+++ score.js
@@ -227,12 +227,13 @@ export const electricity_location = Object.keys(electricity_factor);
 //Return the calculated value of the electricity entry if score updates successfully.
 export async function electricityVal(location, usage) {
     try {
-        let converted_usage = Math.round(usage / 10) / 100;
+        let converted_usage = usage / 1000;
         if(electricity_factor.hasOwnProperty(location)) {
             let emission_factor = electricity_factor[location];
             let val = Math.round(emission_factor * converted_usage  * 0.453*-10);
 
-            const new_score = await updateScore(252 + val);
+            //const new_score = await updateScore(252 + val);
+            const new_score = await updateScore(val);
             
             if(new_score != undefined) {
                 await saveVal(val, "electricity");
